requirements.txt
docopt==0.6.2
dohq-artifactory==0.4.112; python_version < '3.5'
dohq-artifactory>=0.7.377; python_version >= '3.5'
Jinja2<2.11; python_version < '3.5'
Jinja2>=2.11; python_version >= '3.5'
patool==1.12
pyunpack==0.1.2
PyYAML<5.2; python_version < '3.5'
PyYAML>=5.2; python_version >= '3.5'
requests<2.22; python_version < '3.5'
requests>=2.22; python_version >= '3.5'
urllib3==1.24.3
ordered-set==3.1.1; python_version < '3.5'
ordered-set>=4.0.2; python_version >= '3.5'
packaging>=20.4
python-dateutil==2.8.1
addict==2.2.1
tabulate>=0.8.7
parse>=1.15.0
wcmatch
dohq-common